overlay: remove write permissions for group and others
If the `init.qcom.rc` file has write permissions for the group, then
it's not getting read. Since anyway the files are meant to be read-only,
let's remove write permissions for the group and others.

The change on file ownership (we set them to root:root) is probably only
cosmetic, as it does not seem to impact the functionality, but it makes
sense to do it anyway.